✨ feat(grpc): Integrate MedicationResolver for dynamic processing
This commit introduces and integrates the `MedicationResolver` component into the gRPC server architecture, enabling dynamic medication resolution during the processing of match queue items. This enhancement provides a flexible mechanism to identify and enrich medication data based on configurable resolvers.

The primary goal is to allow the system to intelligently resolve medications, potentially improving matching accuracy and data curation workflows by offering an optional, pluggable resolution service. Additionally, numeric score types in the pharma analysis tool have been upgraded for higher precision.

**Changes made:**
-   **Added** `MedicationResolver` as an `Option<Arc<MedicationResolver>>` dependency to the `GrpcDependencies` struct in `core/src/grpc/params.rs`, making it available for gRPC services.
-   **Introduced** a `with_medication_resolver` builder method in `GrpcDependencies` for flexible dependency injection and setup.
-   **Integrated** the `MedicationResolver` into the `PharmaService` struct in `core/src/grpc/server.rs`, ensuring it is accessible within gRPC handlers.
-   **Implemented** dynamic medication resolution logic within the `handle_process_queue_item` function in `core/src/grpc/server.rs`, allowing the resolver to be conditionally used based on its presence.
-   **Updated** numeric score types from `f32` to `f64` in `core/crates/pharma-analysis/src/main.rs` for `score_a`, `score_b`, and `score` variables, providing increased precision when calculating and displaying scores.

> **Note:** The `MedicationResolver` is designed as an optional dependency, allowing the gRPC service to operate with or without this advanced resolution capability based on runtime configuration. This ensures backward compatibility while providing new functionality.

**Technical details:**
-   Files modified:
    -   `core/src/grpc/params.rs`
    -   `core/src/grpc/server.rs`
    -   `core/crates/pharma-analysis/src/main.rs`
-   New fields: `medication_resolver: Option<Arc<MedicationResolver>>` added to both `GrpcDependencies` and `PharmaService` structs.
-   New methods: `GrpcDependencies::with_medication_resolver` for a fluent API to configure the resolver.
-   Type changes: `f32` replaced with `f64` for score-related variables in the `handle_curate_action` function within `pharma-analysis`.

+++ core/crates/pharma-analysis/src/main.rs
@@ -1127,15 +1127,15 @@ async fn handle_curate_action(
 
                 let mut sorted: Vec<_> = combined.into_iter().collect();
                 sorted.sort_by(|a, b| {
-                    let score_a: f32 = a.1.try_get_by_index(3).unwrap_or(0.0);
-                    let score_b: f32 = b.1.try_get_by_index(3).unwrap_or(0.0);
+                    let score_a: f64 = a.1.try_get_by_index(3).unwrap_or(0.0);
+                    let score_b: f64 = b.1.try_get_by_index(3).unwrap_or(0.0);
                     score_b.partial_cmp(&score_a).unwrap()
                 });
 
                 for (id, row) in sorted {
                     let name: String = row.try_get_by_index(1)?;
                     let strength: Option<String> = row.try_get_by_index(2)?;
-                    let score: f32 = row.try_get_by_index(3)?;
+                    let score: f64 = row.try_get_by_index(3)?;
                     let method: String = row.try_get_by_index(4)?;
 
                     table.add_row(row![
